feat(versions): Add conditional versioning to Git tags
I added the ability to use conditional versions for Terraform
registry models in v2.0.0 of xterrafile. I wanted to see if the
same functionality would be possible for git tags. As long as
they follow a semver (examples: 1.0.0 or v1.2.0) pattern this
is now possible.

It's functional and performant but... I'm not 100% happy with
some of the code and will come back to improve on it as my
experiences and understandings grow.